Refactored Dialog to MaterialDialog with theme
