+++ app/src/main/res/values/styles.xml
@@ -48,4 +48,13 @@
         <item name="android:textColor">@color/materialText2</item>
     </style>
 
+    <style name="MaterialLightDialogTheme" parent="Theme.AppCompat.Light.Dialog.Alert">
+        <!--buttons color-->
+        <item name="colorAccent">@color/colorPrimary</item>
+        <!--title and message color-->
+        <item name="android:textColorPrimary">@color/materialText1</item>
+        <item name="android:textColorSecondary">@color/materialText2</item>
+        <!--dialog background-->
+        <item name="android:windowBackground">@android:color/white</item>
+    </style>
 </resources>
